requirements.txt
google-genai
google-generativeai
instagrapi
Pillow
requests
dotenv